[minor] Xml/HtmlResponse::dump($selector) now dumps outer html
