+++ src/Browser/Response/HtmlResponse.php
@@ -3,13 +3,11 @@
 namespace Zenstruck\Browser\Response;
 
 use Symfony\Component\DomCrawler\Crawler;
-use Symfony\Component\VarDumper\VarDumper;
-use Zenstruck\Browser\Response;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-class HtmlResponse extends Response
+class HtmlResponse extends DomResponse
 {
     public function crawler(): Crawler
     {
@@ -17,27 +15,5 @@ class HtmlResponse extends Response
         $crawler->addHtmlContent($this->body());
 
         return $crawler;
-    }
-
-    /**
-     * @internal
-     */
-    final public function dump(?string $selector = null): void
-    {
-        if (null === $selector) {
-            parent::dump();
-
-            return;
-        }
-
-        $elements = $this->crawler()->filter($selector);
-
-        if (0 === $elements->count()) {
-            throw new \RuntimeException("Element \"{$selector}\" not found.");
-        }
-
-        $elements->each(function(Crawler $node) {
-            VarDumper::dump($node->html());
-        });
     }
 }
